scripts changes

+++ Scripts/image_error.py
@@ -10,8 +10,8 @@ if __name__ == "__main__":
     img_file_0 = sys.argv[1]
     img_file_1 = sys.argv[2]
 
-    img_0 = cv2.imread(img_file_0)
-    img_1 = cv2.imread(img_file_1)
+    img_0 = cv2.imread(img_file_0).astype("float32")
+    img_1 = cv2.imread(img_file_1).astype("float32")
 
     error = mse(img_0,img_1)
